sum_of_products

Function parse

pub fn parse<I: Iterator<Item = Result<String>>>(
    lines: I,
) -> Result<Vec<Vec<f64>>, Box<dyn Error>>

Errors

Will return any Err from lines, or an Err if any word cannot be parsed as an f64.